MAINT: Refactor SQS to structure_builders sub-package
