+++ prlworkflows/structure_builders/__init__.py
@@ -1,0 +1,3 @@
+"""Functions to build Structures"""
+
+from .sqs_db import SQSDatabase, get_structures_from_database
